# backend/src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Bitácora SOC API
  description: API completa para el sistema de Bitácora SOC con gestión de entradas, checklist de turno, usuarios y reportes
  version: 1.0.0
  contact:
    name: Bitácora SOC Team
servers:
  - url: http://192.168.100.50:3000/api
    description: Servidor de desarrollo (IP)
  - url: http://localhost:3000/api
    description: Servidor local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Entry:
      type: object
      properties:
        _id:
          type: string
        content:
          type: string
        entryType:
          type: string
          enum: [operativa, incidente]
        entryDate:
          type: string
          format: date
        entryTime:
          type: string
          pattern: '^([01]\d|2[0-3]):([0-5]\d)$'
        tags:
          type: array
          items:
            type: string
        createdBy:
          type: string
        createdByUsername:
          type: string
        isGuestEntry:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /entries:
    get:
      tags:
        - Entradas
      summary: Listar entradas
      description: Obtiene entradas con filtros y paginación
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: search
          schema:
            type: string
          description: Búsqueda de texto completo
        - in: query
          name: tags
          schema:
            type: string
          description: Tags separados por comas
          example: trellix,hunting
        - in: query
          name: entryType
          schema:
            type: string
            enum: [operativa, incidente]
        - in: query
          name: startDate
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
        - in: query
          name: userId
          schema:
            type: string
      responses:
        '200':
          description: Lista de entradas
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/Entry'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer
                      totalPages:
                        type: integer
    
    post:
      tags:
        - Entradas
      summary: Crear entrada
      description: Crea una nueva entrada en la bitácora
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
                - entryType
                - entryDate
                - entryTime
              properties:
                content:
                  type: string
                  example: Revisión de alertas en #Trellix. Todo operativo. #hunting
                entryType:
                  type: string
                  enum: [operativa, incidente]
                  example: operativa
                entryDate:
                  type: string
                  format: date
                  example: "2025-12-17"
                entryTime:
                  type: string
                  example: "14:30"
      responses:
        '201':
          description: Entrada creada
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  entry:
                    $ref: '#/components/schemas/Entry'
  /entries/tags/suggest:
    get:
      tags:
        - Entradas
      summary: Autocompletar tags
      description: Sugiere tags basados en el texto ingresado
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
          example: tre
      responses:
        '200':
          description: Lista de sugerencias
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    tag:
                      type: string
                    count:
                      type: integer
...